// frontend/src/api/rest-client/dto/messageCreateRequestDtoRole.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type MessageCreateRequestDtoRole = typeof MessageCreateRequestDtoRole[keyof typeof MessageCreateRequestDtoRole];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MessageCreateRequestDtoRole = {
  USER: 'USER',
  ASSISTANT: 'ASSISTANT',
} as const;