// frontend/src/api/rest-client/dto/messageResponseDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { MessageResponseDtoRole } from './messageResponseDtoRole';

export interface MessageResponseDto {
  id: number;
  linearChatId: number;
  content: string;
  role: MessageResponseDtoRole;
  createdAt: string;
  updatedAt: string;
}