// frontend/src/api/rest-client/dto/messageResponseDtoRole.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type MessageResponseDtoRole = typeof MessageResponseDtoRole[keyof typeof MessageResponseDtoRole];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MessageResponseDtoRole = {
  USER: 'USER',
  ASSISTANT: 'ASSISTANT',
} as const;