// frontend/src/api/rest-client/dto/linearChatTreeNode.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface LinearChatTreeNode {
  id: number;
  graphChatId: number;
  parentLinearId: number;
  attachedMessageId: number;
  title: string;
  isLoading: boolean;
  createdAt: string;
  children: LinearChatTreeNode[];
}